+++ 27---React-Server-Components-And-Server-Actions-A-Closer-Look/01-starting-project/app/actions/users.js
@@ -1,0 +1,16 @@
+"use server";
+
+import fs from "node:fs";
+
+export async function saveUserAction(formData) {
+  console.log("saveUserAction excuted");
+  const data = fs.readFileSync("dummy-db.json", "utf-8");
+  const instructors = JSON.parse(data);
+  const newInstructor = {
+      id: new Date().getTime().toString(),
+      name: formData.get("name"),
+      title: formData.get("title")
+  };
+  instructors.push(newInstructor);
+  fs.writeFileSync("dummy-db.json", JSON.stringify(instructors));
+}
