+++ 27---React-Server-Components-And-Server-Actions-A-Closer-Look/01-starting-project/app/components/ServerActionDemo.js
@@ -1,21 +1,7 @@
-// "use client" // would result in Error:   × It is not allowed to define inline "use server" annotated Server Actions in Client Components.
+import { saveUserAction } from "@/actions/users";
 
-import fs from "node:fs";
+export default function ServerActionsDemo() {
 
-export default async function ServerActionsDemo() {
-    async function saveUserAction(formData) {
-        "use server";
-        console.log("saveUserAction excuted");
-        const data = fs.readFileSync("dummy-db.json", "utf-8");
-        const instructors = JSON.parse(data);
-        const newInstructor = {
-            id: new Date().getTime().toString(),
-            name: formData.get("name"),
-            title: formData.get("title")
-        };
-        instructors.push(newInstructor);
-        fs.writeFileSync("dummy-db.json", JSON.stringify(instructors));
-    }
     return (
         <div className="rsc">
             <h2>Server Actions</h2>
