+++ 27---React-Server-Components-And-Server-Actions-A-Closer-Look/01-starting-project/dummy-db.json
@@ -1,1 +1,1 @@
-[{"id":"u1","name":"Max","title":"instructor"},{"id":"u2","name":"Manuel","title":"instructor"},{"id":"1734888103974","name":"Fred","title":"Lewitt"}]
+[{"id":"u1","name":"Max","title":"instructor"},{"id":"u2","name":"Manuel","title":"instructor"},{"id":"1734888103974","name":"Fred","title":"Lewitt"},{"id":"1734888652664","name":"Fred (again)","title":"Lewitt (Again)"},{"id":"1734888683510","name":"Hugo (again)","title":"Machefer (again)"}]
